# docs/api/openapi.yaml
...
openapi: 3.0.3
info:
  title: 优刻（YOUC）平台 API
  version: 0.19.0-unreleased
  description: |
    优刻全站接口：用户面统一前缀 `/api/v1`，管理面为 `/api/admin`（不进公开版本契约）。

    鉴权：
    - 业务接口走 `Authorization: Bearer <accessToken>`（HS256 JWT，15 分钟有效，登录/刷新接口下发）。
    - 刷新令牌走 HttpOnly cookie `ic_refresh`（Path=/api，30 天，单次轮换，服务端只存哈希）。
    - 媒体读取额外接受 HttpOnly cookie `ic_media`（Path=/api/v1/media，为 `<img>/<video>` 直链准备的只读凭据；
      `/api/v1/media` 与 `/api/v1/media-download` 的 GET/HEAD 可用 Bearer 或该 cookie，写操作仅认 Bearer）。

    全局约定：
    - 统一错误信封 `{"error":{"code":"...","message":"...","fields":{...}}}`（fields 仅字段级错误时出现，
      error 对象可能平铺附加键，如 REVISION_CONFLICT 附 revision）；限流响应附 `Retry-After` 头。
    - 时间一律 RFC3339Nano UTC 字符串；金额单位为微元（1 元 = 1000000 micros）。
    - 分页：page/size（page 从 1 起、size 默认 20 上限 100）或游标（cursor + nextCursor）两种，按接口标注。
    - 站点维护模式开启时，非管理员写请求一律 503 MAINTENANCE_MODE（认证、OIDC、管理面、支付回调豁免）。
    - 强制改密账号仅可调用登出、刷新与修改密码，其余接口 403 PASSWORD_CHANGE_REQUIRED。
    - 通用错误（除特别标注外各组适用）：401 UNAUTHORIZED / TOKEN_EXPIRED、403 ACCOUNT_DISABLED、
      403 PASSWORD_CHANGE_REQUIRED、503 MAINTENANCE_MODE、500 INTERNAL_ERROR。
servers:
  - url: https://sim-art.youc.online
    description: 测试环境
  - url: https://youc.online
    description: 正式环境（主站应用部署后同域；当前该域为静态官网）
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 登录/刷新接口返回的 accessToken（HS256，15 分钟）。
    mediaCookie:
      type: apiKey
      in: cookie
      name: ic_media
      description: 媒体只读 cookie（Path=/api/v1/media，HttpOnly），登录/刷新时下发轮换；仅 GET/HEAD 可代替 Bearer。
  parameters:
    PageParam:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
      description: 页码，从 1 开始。
    SizeParam:
      name: size
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      description: 每页数量，1-100，默认 20；越界 400 VALIDATION_FAILED。
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          required: [code, message]
          additionalProperties: true
          properties:
            code: { type: string, description: 错误码，如 VALIDATION_FAILED }
            message: { type: string, description: 中文错误消息 }
            fields:
              type: object
              additionalProperties: { type: string }
              description: 字段级错误说明（键为字段名）。
...
    ModerationRecord:
      type: object
      description: 审核记录（只存内容哈希与风险标签，不存原始提示词）。
      properties:
        id: { type: string, format: uuid }
        userId: { type: string, format: uuid }
        stage: { type: string, enum: [upload, artifact], description: upload=上传送审，artifact=生成产物送审 }
        contentType: { type: string }
        contentHash: { type: string }
        provider: { type: string }
        providerRequestId: { type: string, nullable: true }
        policyVersion: { type: string }
        decision: { type: string, enum: [passed, rejected, error] }
        riskLabels: { type: array, items: { type: string } }
        reviewStatus: { type: string, enum: [pending, approved, rejected] }
        reviewRevision: { type: integer, description: 人工改判乐观锁 }
        reviewNote: { type: string, nullable: true }
        compensatedMicros: { type: integer, format: int64 }
        createdAt: { type: string }
        generationId: { type: string, description: 有则返回 }
        reviewedBy: { type: string, description: 有则返回 }
        reviewedAt: { type: string, description: 有则返回 }
        quarantineExpiresAt: { type: string, description: 有则返回 }
...
  responses:
...
    Error403:
      description: 无权限或账号状态受限（FORBIDDEN / ACCOUNT_DISABLED / EMAIL_NOT_VERIFIED / READ_ONLY 等）
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
paths:
...
  /api/admin/moderation/records:
    get:
      tags: [Admin-审核]
      summary: 审核记录筛选（moderation.read）
      parameters:
        - { $ref: '#/components/parameters/PageParam' }
        - { $ref: '#/components/parameters/SizeParam' }
        - { name: stage, in: query, schema: { type: string, enum: [upload, artifact] } }
        - { name: decision, in: query, schema: { type: string, enum: [passed, rejected, error] } }
        - { name: reviewStatus, in: query, schema: { type: string, enum: [pending, approved, rejected] } }
        - { name: userId, in: query, schema: { type: string, format: uuid } }
        - { name: label, in: query, schema: { type: string }, description: 风险标签包含匹配 }
        - { name: sort, in: query, schema: { type: string }, description: '白名单 createdAt|decision，默认 -createdAt' }
      responses:
        '200':
          description: 审核记录分页
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/ModerationRecord' } }
                  total: { type: integer, format: int64 }
                  page: { type: integer }
                  size: { type: integer }
        '403': { $ref: '#/components/responses/Error403' }
...